Split config and logic
